Toaster added and formatting

+++ src/app/Component/student-list/student-list.component.ts
@@ -1,47 +1,55 @@
 import { Component, OnInit } from '@angular/core';
-import { StudentService } from '../../Services/student.service';
+import { StudentService } from '../../Services/Student/student.service';
 import { NgFor, NgIf } from '@angular/common';
 import { HeaderComponent } from '../header/header.component';
+import { NotificationService } from '../../Services/Notification/nitification.service';
 
 @Component({
   selector: 'app-student-list',
-  imports: [NgFor,HeaderComponent,NgIf],
+  imports: [NgFor, HeaderComponent, NgIf],
   templateUrl: './student-list.component.html',
-  styleUrl: './student-list.component.css'
+  styleUrl: './student-list.component.css',
 })
-export class StudentListComponent implements OnInit{
-
-  studentList : any[] = [];
+export class StudentListComponent implements OnInit {
+  studentList: any[] = [];
   isLoading: boolean = false;
 
-  constructor(private studentService: StudentService){}
+  constructor(
+    private studentService: StudentService,
+    private notificationService: NotificationService
+  ) {}
 
   ngOnInit(): void {
     this.getAllStudents();
   }
 
-  getAllStudents(){
+  getAllStudents() {
     this.isLoading = true;
-    this.studentService.getStudents().subscribe((response:any)=>{
-      console.log("student list is",response);
+    this.studentService.getStudents().subscribe((response: any) => {
       this.studentList = response;
       this.isLoading = false;
-      console.log("student list after assign",this.studentList);
-    })
+    });
     setTimeout(() => {
       this.isLoading = false;
     }, 15000);
   }
 
-  deleteStudent(id:number){
-    this.studentService.deleteStudent(id).subscribe((response:any)=>{
-      console.log("deleted response is",response);
-      if(response){
-        alert("student data deleted successfully");
-        this.getAllStudents();
-      }else{
-        alert("failed to delete student data");
-      }
-    })
+  deleteStudent(id: number) {
+    if (confirm('Are you sure to delete this student data?')) {
+      this.studentService.deleteStudent(id).subscribe((response: any) => {
+        if (response) {
+          this.notificationService.showSuccess(
+            'Student data deleted successfully',
+            'Success'
+          );
+          this.getAllStudents();
+        } else {
+          this.notificationService.showError(
+            'Failed to delete student data',
+            'Error'
+          );
+        }
+      });
+    }
   }
 }
